LayoutTests/platform/ios-simulator/editing/execCommand/break-out-of-empty-list-item-expected.txt

This tests CompositeEditCommand::breakOutOfEmptyListItem by inserting new paragraph

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL enterAtTarget('<ul><li>a <ul><li>b</li><li id="target"></li></ul> </li></ul>') should be <ul><li>a </li><ul><li>b</li></ul><li><br></li> </ul>. Was <ul><li>a <ul><li>b</li><li id="target"></li></ul> </li></ul>.
FAIL enterAtTarget('<ul><li>a <ul><li id="target"></li><li>b</li></ul> </li></ul>') should be <ul><li>a </li><li><br></li><ul><li>b</li></ul> </ul>. Was <ul><li>a <ul><li id="target"></li><li>b</li></ul> </li></ul>.
FAIL enterAtTarget('<ul><li>a <ul><li>b</li><li id="target"></li><li>c</li></ul> </li></ul>') should be <ul><li>a </li><ul><li>b</li></ul><li><br></li><ul><li>c</li></ul> </ul>. Was <ul><li>a <ul><li>b</li><li id="target"></li><li>c</li></ul> </li></ul>.
FAIL enterAtTarget('<ul><li>hello<ul><li id="target"><br></li></ul>world</li></ul>') should be <ul><li>hello<div><br></div>world</li></ul>. Was <ul><li>hello<ul><li id="target"><br></li></ul>world</li></ul>.
FAIL enterAtTarget('<ul><li>hello<ul><li id="target"><br></li></ul></li></ul>') should be <ul><li>hello</li><li><br></li></ul>. Was <ul><li>hello<ul><li id="target"><br></li></ul></li></ul>.
FAIL enterAtTarget('<ul><li><ul><li id="target"><br></li></ul>world</li></ul>') should be <ul><li><div><br></div>world</li></ul>. Was <ul><li><ul><li id="target"><br></li></ul>world</li></ul>.
FAIL enterAtTarget('<ul><li><ul><li id="target"><br></li></ul></li></ul>') should be <ul><li></li><li><br></li></ul>. Was <ul><li><ul><li id="target"><br></li></ul></li></ul>.
FAIL enterAtTarget('<ul><li>hello</li><br id="target"></ul>') should be <ul><li>hello</li></ul><div><br></div>. Was <ul><li>hello</li><br id="target"></ul>.
FAIL enterAtTarget('<ul><br id="target"></ul>') should be <div><br></div>. Was <ul><br id="target"></ul>.
FAIL enterAtTarget('<ul><li>hello</li>abc<li id="target"></li></ul>') should be <ul><li>hello</li>abc</ul><div><br></div>. Was <ul><li>hello</li>abc<li id="target"></li></ul>.
FAIL enterAtTarget('<ul><li>1</li><ul><li>2.1</li></ul><li id="target"></li></ul>') should be <ul><li>1</li><ul><li>2.1</li></ul></ul><div><br></div>. Was <ul><li>1</li><ul><li>2.1</li></ul><li id="target"></li></ul>.
FAIL enterAtTarget('<ul><li>1</li><ul><li>2.1</li><li>2.2</li><li id="target"></li></ul><li>3</li></ul>') should be <ul><li>1</li><ul><li>2.1</li><li>2.2</li></ul><li><br></li><li>3</li></ul>. Was <ul><li>1</li><ul><li>2.1</li><li>2.2</li><li id="target"></li></ul><li>3</li></ul>.
FAIL enterAtTarget('<ul><li>1</li><ul><li>2.1</li><li>2.2</li>abc<li id="target"></li></ul><li>3</li></ul>') should be <ul><li>1</li><ul><li>2.1</li><li>2.2</li>abc</ul><li><br></li><li>3</li></ul>. Was <ul><li>1</li><ul><li>2.1</li><li>2.2</li>abc<li id="target"></li></ul><li>3</li></ul>.
FAIL enterAtTarget('<ul><li>1</li><li id="target"></li><li>3</li></ul>') should be <ul><li>1</li></ul><div><br></div><ul><li>3</li></ul>. Was <ul><li>1</li><li id="target"></li><li>3</li></ul>.
PASS successfullyParsed is true

TEST COMPLETE
